Server/requirements.txt
djangorestframework==3.14.0
pillow==10.0.0
django-rest-authtoken==2.1.4
drf-yasg==1.21.7